Create file for system Four Biharmonic Phase Oscillators
